Add files via upload
This is my first project with PowerBI. I wanted to learn and practice data visualization techniques. The data used in this project is entirely hypothetical and was generated using AI to simulate a retail store’s sales data. It does not reflect any real-world organization or business. 
I didn't have a proper database to practice so I asked AI for some data to get a headstart with the project. 
I aimed to build this dashboard to:
1. Learn how to clean and transform data.
2. Create visualizations such as bar charts and KPIs to showcase sales performance.
3. Use slicers and filters to interact with the data.
Through this project, I have explored key Power BI features like slicers and data transformation. 
Since I am a beginner at this, I would appreciate any constructive criticism on the project.